backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt
sqlalchemy
pydantic
pydantic-settings
email-validator
openai
openai-whisper
opencv-python
numpy
Pillow
httpx
aiofiles
python-dotenv
azure-cognitiveservices-speech
azure-cognitiveservices-vision-face
msrest
scikit-learn
torch
torchvision
librosa
soundfile
azure-cognitiveservices-vision-face

